Move `runtime.py` to virool workflow.
rename `virtool_workflow.runtime.py` to `environment.py`.
